simplified test
