+++ Regression/WarpX-tests.ini
@@ -1010,8 +1010,7 @@ doVis = 0
 [LaserInjectionFromTXYEFile]
 buildDir = .
 inputFile = Examples/Tests/laser_injection_from_file/test_script.py
-aux1File = Examples/Tests/laser_injection_from_file/test_injection.py
-aux2File = Examples/Tests/laser_injection_from_file/inputs.2d_test_txye
+aux1File = Examples/Tests/laser_injection_from_file/inputs.2d_test_txye
 customRunCmd = ./test_script.py
 dim = 2
 addToCompileString =
